pyproject.toml
[tool.poetry]
name = "tracker api"
version = "0.1.0"
description = ""
authors = ["Muhammed Uluel <mhmmduluel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
django-timescaledb = "^0.2.13"
environs = "^11.2.0"
psycopg2-binary = "^2.9.10"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.27.2"
markdown = "^3.7"
django-filter = "^24.3"
gunicorn = "^23.0.0"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
